+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,14 @@
         android:theme="@style/Theme.CarnivalTrip">
         <activity
             android:name=".MainActivity"
-            android:exported="true" />
-        <activity
-            android:name=".MainPoi"
-            android:exported="true"></activity>
+            android:exported="true" >
+
+        </activity>
         <activity
             android:name=".DetailPoi"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
