+++ app/src/main/java/com/example/carnivaltrip/DetailPoi.kt
@@ -2,7 +2,6 @@ package com.example.carnivaltrip
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_main_poi.*
 
 class DetailPoi : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
