+++ app/src/main/java/com/example/carnivaltrip/MainActivity.kt
@@ -2,6 +2,11 @@ package com.example.carnivaltrip
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.gson.GsonBuilder
+import kotlinx.android.synthetic.main.fragment_view_list_poi.*
+import java.io.IOException
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -13,4 +18,6 @@ class MainActivity : AppCompatActivity() {
             .replace(R.id.content, ViewListPoi())
             .commit()
     }
+
+
 }
