--- app/src/main/res/layout/activity_main_poi.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/content"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="15dp"
-    tools:context=".MainPoi">
-
-    <TextView
-        android:id="@+id/title_activity"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:text="@string/title_list_POI"
-        android:textSize="34sp"
-        style="@style/TextAppearance.AppCompat.Title">
-
-    </TextView>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/r_list_poi"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="20dp"
-        android:paddingBottom="50dp"
-        app:layout_constraintTop_toBottomOf="@+id/title_activity"
-        tools:layout_editor_absoluteX="16dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
